backend/requirements.txt
fastapi
uvicorn
supabase
groq
python-multipart
python-dotenv
pydantic
httpx
